update to TeX Live 2017
Also add "biber" package, and automate a little more on how we choose font
packages. Whereas before we had a whitelist of font packages, we now instead
grab a larger set to begin with and then whittle it down with a smaller set of
blacklist patterns.

We also add a new image type under `texlive-with-fonts-and-user'.