texlive-with-fonts-and-user: add git
This is useful for TeX documents that make use of git-based version control. The
idea is that we need the git binary inside the Docker image, and then when we
run the container by mounting a host filesystem project folder that is tracked
by git, the container's git can read that information (and inject it into the
TeX document).