Install dependencies for `latexindent`
I'm not sure if this is something you want in the repo, but I added it to my branch. 

It installs the dependencies for `latexindent` (after the path so that `cpan` is found). 
The `~/.cpan` directory is then nuked because its no longer necessary. I'm unsure if there is other stuff that can be removed, though.