add helper scripts for increased automation
These scripts are a little messy, but they are better than nothing. We
should probably consolidate to a single script which can take arguments
and such, but this will do for now.

Also, make texlive-user completely tailored to the uid/gid of the
end-user, to make running it as seamless as possible.