travis: use newer version of Ubuntu
This gives us a newer version of Docker, because the 18.06 version used
in Xenial (current default) is too ancient. It also runs into issues
with the /etc folder (which breaks mktexlsr and other latex-* package
install hooks in Arch Linux from within the Docker container) that is
not reproducible in the newer 19.03 version.

Fixes #6.